Refactor: move train method to vw_base
